Updating the lock_free_buffer_queue to better match the proposed queue spec where failed copy operations don't insert and always pop.
Also updates a lot of the atomic logic to be a little less reliant
on atomics and with more relaxed fencing. This also moves the burden of
the fencing on the state of the elements in the queue rather than
on the presence of data, which means we have a single update which
does the commit (linearizable queue).

Still not wait-free though.